Attach course info login link
On mobile Firefox, the new tab will close immediately if the user has
not logged in. It's also not easy to find the login button on the Online
Course Information website. Attach a link in the form to make it easier
for users to log in. This will help ensure that users can quickly access
the login page without having to navigate through the website.